Modification to the ball movement
Modified the ball movement, so it no longer only moves at a velocity of (1,1).
By calculating the ball velocity as a unit vector, it makes the ball move at the same speed no matter the direction.

Also changed the way the ball reacts to collision with the player, to add a bit more variation to the movement.